Clean up private variables and move get_local_addresses to utils
