+++ dzmq/utils.py
@@ -1,5 +1,8 @@
 import logging
 import base64
+import socket
+
+import netifaces
 try:
     import ujson as json
 except ImportError:
@@ -115,3 +118,62 @@ def pack_msg(obj):
             obj[key] = pack_msg(value)
 
     return json.dumps(obj).encode('utf-8')
+
+
+# Taken from rosgraph
+# https://github.com/ros/ros_comm/blob/hydro-devel/tools/rosgraph/src/rosgraph/network.py
+# cache for performance reasons
+_local_addrs = None
+
+
+def get_local_addresses(use_ipv6=False, addrs=None, ifaces=None):
+    """
+    Get a list of local ip addresses that meet a given criteria.
+
+    Parameters
+    ----------
+    use_ipv6 : bool, optional
+        Whether to allow ipv6 addresses.
+    addrs : list of strings, optional
+        List of addresses to look for.
+    ifaces : list strings, optional
+        List of ethernet interfaces.
+
+    Returns
+    -------
+    address : list of strings
+        List of available local ip addresses that meet a given criteria.
+    """
+    # cache address data as it can be slow to calculate
+    global _local_addrs
+    if _local_addrs is not None:
+        return _local_addrs
+
+    ifaces = ifaces or netifaces.interfaces()
+
+    v4addrs = []
+    v6addrs = []
+    for iface in ifaces:
+        try:
+            ifaddrs = netifaces.ifaddresses(iface)
+        except ValueError:
+            # even if interfaces() returns an interface name
+            # ifaddresses() might raise a ValueError
+            # https://bugs.launchpad.net/ubuntu/+source/netifaces/+bug/753009
+            continue
+        if socket.AF_INET in ifaddrs:
+            v4addrs.extend([addr
+                            for addr in ifaddrs[socket.AF_INET]
+                            if 'broadcast' in addr])
+        if socket.AF_INET6 in ifaddrs:
+            v6addrs.extend([addr
+                            for addr in ifaddrs[socket.AF_INET6]
+                            if 'broadcast' in addr])
+    if use_ipv6:
+        local_addrs = v6addrs + v4addrs
+    else:
+        local_addrs = v4addrs
+    if addrs:
+        local_addrs = [a for a in local_addrs if not a['addr'] in addrs]
+    _local_addrs = local_addrs
+    return local_addrs
